requirements.txt
appnope
asttokens
backcall
beautifulsoup4
blinker
Bottleneck
Brotli
certifi
cffi
charset-normalizer
click
comm
cryptography
debugpy
decorator
dnspython
executing
Flask
Flask-Cors
gunicorn
idna
importlib-metadata
ipykernel
ipython
itsdangerous
jedi
Jinja2
jupyter_client
jupyter_core
MarkupSafe
matplotlib-inline
nest_asyncio
numexpr
numpy
packaging
pandas
parso
pexpect
pickleshare
platformdirs
prompt_toolkit
psutil
ptyprocess
pure-eval
pycparser
Pygments
pymongo
PySocks
python-dateutil
python-dotenv
pytz
pyzmq
requests
six
soupsieve
stack-data
torch
torchio
torchvision
tornado
traitlets
transformers
typing_extensions
tzdata
urllib3
wcwidth
Werkzeug
zipp
